minor fix for saving

+++ homeassistant/setup.sh
@@ -1,4 +1,14 @@
 sleep 3
 
-cp /tmp/configuration.yaml /config/configuration.yaml
-cp /tmp/intents.yaml /config/intents.yaml 
+if [ -e /config/configuration.yaml ]; then
+    echo "config file already exists";
+else
+    cp /tmp/configuration.yaml /config/configuration.yaml
+fi
+
+if [ -e /config/intents.yaml ]; then
+    echo "intents file already exists";
+else
+    cp /tmp/intents.yaml /config/intents.yaml;
+fi
+
